TURMERIC-1555: Clean various unit tests setups.
This cleans up various unit test setups so they acctually make it to their
corresponding unit test.  Along the way this fixed the LockAssetTest so that
it now worksk.  I suspect there is an issue with missing meta data before that
caused it to fail.  Will need to add more validation checks in the
receiving methods to make sure required data is there.