Retrieve the error message when a test fails.
Currently most of the test when they fail don't return a detailed error
message as to why.  This means you have to do more debugging then you should
when doing unit testing.   Created a generic method that can take
a response object and return the first error message that is found.  If
no error message is return then it returns null.