TURMERIC-1607: refactor the updateAssetAttributes to a Strategy pattern
Much of the code in RepositoryServiceProviderImpl is implemented in line, the problem
is that the class itself is doing too much.  In order to make it easier to maintain
the Strategy pattern is implemented to break the functionality out.  It is similar in
some ways to Proxy pattern, but the Strategy actually implements the code instead
of calls another class.   The RepositoryProviderImpl pretty much becomes the
Proxy pattern once the refactoring is done.